Labs/GCP_Labs/Cloud_Runner_Labs/Intermediate_Lab/cloud-run-intermediate-app/requirements.txt
flask
gunicorn
google-cloud-storage
google-cloud-bigquery
